Moves rememberProcessedText from StyleFactory to TextComponentView.
